+++ protoc_plugin/lib/file_generator.dart
@@ -117,12 +117,12 @@ class FileGenerator extends ProtobufContainer {
   /// Used to avoid collisions after names have been mangled to match the Dart
   /// style.
   final Set<String> usedTopLevelNames = Set<String>()
-    ..addAll(toplevelReservedCapitalizedNames);
+    ..addAll(forbiddenTopLevelNames);
 
   /// Used to avoid collisions in the service file after names have been mangled
   /// to match the dart style.
   final Set<String> usedTopLevelServiceNames = Set<String>()
-    ..addAll(toplevelReservedCapitalizedNames);
+    ..addAll(forbiddenTopLevelNames);
 
   final Set<String> usedExtensionNames = Set<String>()
     ..addAll(forbiddenExtensionNames);