+++ protoc_plugin/lib/names.dart
@@ -447,13 +447,12 @@ bool _isDartFieldName(name) => name.startsWith(_dartFieldNameExpr);
 
 final _dartFieldNameExpr = RegExp(r'^[a-z]\w+$');
 
-/// Names that would collide with capitalized core Dart names as top-level
-/// identifiers.
-final List<String> toplevelReservedCapitalizedNames = const <String>[
+/// Names that would collide as top-level identifiers.
+final List<String> forbiddenTopLevelNames = <String>[
   'List',
   'Function',
   'Map',
-];
+]..addAll(_dartReservedWords);
 
 final List<String> reservedMemberNames = <String>[]
   ..addAll(_dartReservedWords)